+++ src/pages/HomeBottomBanners/editHomeBottomBanner.js
@@ -77,15 +77,17 @@ const EditBanner = () => {
   useEffect(() => {
     context.setProgress(20);
     fetchDataFromApi("/api/imageUpload").then((res) => {
-      res?.map((item) => {
-        item?.images?.map((img) => {
-          deleteImages(`/api/homeBottomBanners/deleteImage?img=${img}`).then(
-            (res) => {
-              deleteData("/api/imageUpload/deleteAllImages");
-            }
-          );
+      if (Array.isArray(res)) {
+        res?.map((item) => {
+          item?.images?.map((img) => {
+            deleteImages(`/api/homeBottomBanners/deleteImage?img=${img}`).then(
+              (res) => {
+                deleteData("/api/imageUpload/deleteAllImages");
+              }
+            );
+          });
         });
-      });
+      }
     });
 
     fetchDataFromApi(`/api/homeBottomBanners/${id}`).then((res) => {
@@ -178,15 +180,19 @@ const EditBanner = () => {
           setTimeout(() => {
             setUploading(false);
             img_arr = [];
-            uniqueArray=[];
+            uniqueArray = [];
             fetchDataFromApi("/api/imageUpload").then((res) => {
+             if (Array.isArray(res)){
               res?.map((item) => {
                 item?.images?.map((img) => {
-                  deleteImages(`/api/homeBottomBanners/deleteImage?img=${img}`).then((res) => {
-                   // deleteData("/api/imageUpload/deleteAllImages");
+                  deleteImages(
+                    `/api/homeBottomBanners/deleteImage?img=${img}`
+                  ).then((res) => {
+                    // deleteData("/api/imageUpload/deleteAllImages");
                   });
                 });
               });
+             }
             });
             context.setAlertBox({
               open: true,
@@ -409,7 +415,6 @@ const EditBanner = () => {
                         <>
                           <input
                             type="file"
-                            
                             onChange={(e) =>
                               onChangeFile(e, "/api/homeBanner/upload")
                             }
